AI-Medical-Assistant/requirements.txt
transformers[torch]
datasets
pandas
scikit-learn
openpyxl
arabic-reshaper
python-bidi
torch
petf
matplotlib